requirements.txt
# MicDysphagiaFramework 依賴列表
# 最後更新日期: 2023-07-27

# 核心依賴
numpy>=1.19.5,<2.1.0     # 2.0.2+ 推薦用於 Apple Silicon
scipy>=1.7.1,<1.14.0      
pyyaml>=6.0,<7.0
tqdm>=4.62.3,<5.0.0
scikit-learn>=1.0.2,<1.7.0
pandas>=1.3.5,<2.3.0
matplotlib>=3.5.1,<4.0.0
seaborn>=0.11.2,<0.14.0

# 音頻處理
librosa>=0.9.1,<0.12.0    # 音頻處理和特徵提取
soundfile>=0.10.3,<0.14.0 # 讀取/寫入音頻文件
torchaudio>=2.0.0,<2.7.0  # PyTorch 音頻庫，用於特徵提取
audiomentations>=0.27.0,<0.40.0  # 音頻數據增強 (可選)

# PyTorch (可透過不同方式安裝)
# macOS 用戶: pip install torch torchvision torchaudio
# CUDA 用戶: conda install pytorch torchvision torchaudio pytorch-cuda=11.8 -c pytorch -c nvidia
torch>=2.0.0,<2.7.0
torchvision>=0.15.0,<0.22.0
timm>=0.9.2,<1.1.0        # Swin Transformer 模型庫
einops>=0.6.0,<0.9.0      # 用於更簡潔地處理張量操作
torchmetrics>=0.11.0,<1.8.0 # 用於計算各種指標
pytorch-lightning>=2.0.0,<2.6.0 # 用於簡化 PyTorch 訓練流程 (可選)

# 視覺化
tensorboard>=2.11.0,<2.20.0  # 訓練可視化工具

# 文件處理
openpyxl>=3.0.10,<3.3.0      # 處理Excel文件
xlrd>=2.0.1,<3.0.0           # 處理舊式Excel文件

# 實用工具
python-dotenv>=0.21.0,<1.2.0  # 環境變量管理
joblib>=1.1.0,<1.5.0         # 並行計算
pillow>=9.3.0,<12.0.0        # 圖像處理
wandb>=0.15.0,<0.20.0        # 實驗追踪 (可選)

# 以下為開發相關依賴（可選）
pytest>=7.2.0        # 單元測試
# black>=22.10.0       # 代碼格式化
# flake8>=6.0.0        # 代碼檢查
# jupyter>=1.0.0       # 筆記本開發 

# 環境設置說明
# 建議使用 conda 環境:
# conda create -n micdys python=3.9
# conda activate micdys
# pip install -r requirements.txt 